+++ src/aks-preview/azext_aks_preview/commands.py
@@ -314,6 +314,13 @@ def load_command_table(self, _):
         g.custom_command("delete", "aks_agentpool_delete", supports_no_wait=True)
         g.custom_command("get-upgrades", "aks_agentpool_get_upgrade_profile")
         g.custom_command(
+            "get-bootstrap-data",
+            "aks_agentpool_get_bootstrap_data",
+            sensitive_info=g.sensitive(
+                sensitive_keys=["bootstrapToken", "caCertData"]
+            ),
+        )
+        g.custom_command(
             "get-rollback-versions",
             "aks_agentpool_get_rollback_versions",
             table_transformer=aks_agentpool_rollback_versions_table_format